Avoid need for possessive noun conjugation and switch out word for pig
Why didn't I catch this sooner? Croatian possessive nouns need to be conjugated (eg. for Katarina: Katarinin/a). Structuring this as "Kineski horoskopski znak osobe "Katarina": prase" is kinda clunky, but is grammatically correct and easier to understand. Also, pigs as a species are more commonly referred to as "svinja" than "prase", out of at least 3 other words for the same animal.
See? Croatian doesn't lack basic words! Just the ones that are important nowadays! 😥